break out utils file
